Removed weight from Entity (weight is redundant as it can be calculated from radius)

+++ core/src/com/game/agar/entities/Player.java
@@ -35,8 +35,7 @@ public class Player extends Entity {
     }
 
     public void eat (Entity collidingObject){
-        int massGained = collidingObject.getWeight();
-
+        float massGained = collidingObject.getWeight();
         updateRadius(massGained);
     }
